Revamp operator controller runtime tests
This change should put us in a better place once we start moving parts
of Contour and the Operator around.

We can lean more heavily on this test suite rather than the e2e test
suite since it should be faster to run/write etc.

This code probably duplicates some things from the e2e suite and things
from that suite could be brought into here to run faster and with less
flakes. Those should be future changes after this starting point.

- Moved from Ginkgo to regular go tests (no need for extra ginkgo
complexity yet)
- Ports existing tests
- Add additional test coverage
- Some small changes to how controller logs handling of namespace
deletion (was confusing to see that the namespace was deleted when it
wasnt actually b/c remove on delete was false)

Signed-off-by: Sunjay Bhatia <sunjayb@vmware.com>